VoxelBase is parent for pixel/super/mega, rename VoxelGroups, VoxelFunctions (#51)
* WIP: VoxelGrouping actions

* Define VoxelBaseClass as parent for pixel/super/mega objects, refactor

* Rename base class to VoxelBase, functions to VoxelFunctions

* WIP: regression. program crashing.

* Fix: VoxelBase is parent for pixel/super/mega, rename VoxelGroups, VoxelFunctions

* Fix: Crash fix. Circular reference.
