--- Phindr3D-Python/src/VoxelGrouping/__init__.py
@@ -1,4 +1,0 @@
-import sys, os.path
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
-
-from .VoxelGrouping import *
