+++ Phindr3D-Python/src/VoxelGroups/VoxelFunctions.py
@@ -1,0 +1,37 @@
+# Copyright (C) 2022 Sunnybrook Research Institute
+# This file is part of src <https://github.com/DWALab/Phindr3D>.
+#
+# src is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# src is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with src.  If not, see <http://www.gnu.org/licenses/>.
+
+class VoxelFunctions:
+    """Static methods for finding voxel properties. Referenced from
+    https://github.com/DWALab/Phindr3D/tree/9b95aebbd2a62c41d3c87a36f1122a78a21019c8/Lib
+    and
+    https://github.com/SRI-RSST/Phindr3D-python/blob/ba588bc925ef72c72103738d17ea922d20771064/phindr_functions.py
+    No constructor. All parameters and methods are static.
+    """
+
+    #@staticmethod
+    ## Main function for returning bin centers of pixels, supervoxels, and mega voxels
+    ## x - m x n (m is number of observations, n is number of channels/category fractions
+    ## numBins - number of categories
+    #def getPixelBins(x, numBins):
+    #    pass
+
+    def temporary(x, numBins):
+        pass
+
+# end VoxelFunctions
+
+
