+++ Phindr3D-Python/src/VoxelGroups/VoxelGroups.py
@@ -1,0 +1,58 @@
+# Copyright (C) 2022 Sunnybrook Research Institute
+# This file is part of src <https://github.com/DWALab/Phindr3D>.
+#
+# src is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# src is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with src.  If not, see <http://www.gnu.org/licenses/>.
+
+try:
+    from .VoxelBase import *
+    from .PixelImage import *
+    from .SuperVoxelImage import *
+    from .MegaVoxelImage import *
+except ImportError:
+    from VoxelBase import *
+    from PixelImage import *
+    from SuperVoxelImage import *
+    from MegaVoxelImage import *
+
+try:
+    from ..PhindConfig.PhindConfig import *
+except ImportError:
+    from src.PhindConfig.PhindConfig import *
+
+class VoxelGroups:
+    """From pixels to supervoxels to megavoxels"""
+
+    def __init__(self):
+        """Constructor"""
+        pass
+
+    def action(self):
+        print("Running the VoxelGroups action method")
+
+
+# end class VoxelGroups
+
+
+
+
+if __name__ == '__main__':
+    """Unit testing"""
+
+    pass
+
+
+
+
+
+# end main
