+++ Phindr3D-Python/src/VoxelGroups/__init__.py
@@ -1,0 +1,9 @@
+import sys, os.path
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+
+from .VoxelGroups import *
+from .VoxelFunctions import *
+from .VoxelBase import *
+from .PixelImage import *
+from .SuperVoxelImage import *
+from .MegaVoxelImage import *
