+++ Phindr3D-Python/src/__init__.py
@@ -7,5 +7,5 @@ from .GUI import *
 from .PhindConfig import *
 from .Segmentation import *
 from .Training import *
-from .VoxelGrouping import *
+from .VoxelGroups import *
